chore(cmd/immuadmin): improve code organization and help messages
