+++ cmd/immuadmin/commands/init.go
@@ -20,6 +20,7 @@ import (
 	c "github.com/codenotary/immudb/cmd"
 	"github.com/codenotary/immudb/cmd/docs/man"
 	"github.com/codenotary/immudb/pkg/client"
+	"github.com/codenotary/immudb/pkg/gw"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -31,7 +32,10 @@ type commandline struct {
 
 type commandlineDisc struct{}
 
-func Init(cmd *cobra.Command) {
+func Init(cmd *cobra.Command, o *c.Options) {
+	if err := configureOptions(cmd, o); err != nil {
+		c.QuitToStdErr(err)
+	}
 	cl := new(commandline)
 	cl.passwordReader = c.DefaultPasswordReader
 
@@ -83,4 +87,45 @@ func (cl *commandline) connect(cmd *cobra.Command, args []string) (err error) {
 		c.QuitToStdErr(err)
 	}
 	return
+}
+
+func configureOptions(cmd *cobra.Command, o *c.Options) error {
+	cmd.PersistentFlags().IntP("port", "p", gw.DefaultOptions().ImmudbPort, "immudb port number")
+	cmd.PersistentFlags().StringP("address", "a", gw.DefaultOptions().ImmudbAddress, "immudb host address")
+	cmd.PersistentFlags().StringVar(&o.CfgFn, "config", "", "config file (default path are configs or $HOME. Default filename is immuadmin.ini)")
+	cmd.PersistentFlags().BoolP("mtls", "m", client.DefaultOptions().MTLs, "enable mutual tls")
+	cmd.PersistentFlags().String("servername", client.DefaultMTLsOptions().Servername, "used to verify the hostname on the returned certificates")
+	cmd.PersistentFlags().String("certificate", client.DefaultMTLsOptions().Certificate, "server certificate file path")
+	cmd.PersistentFlags().String("pkey", client.DefaultMTLsOptions().Pkey, "server private key path")
+	cmd.PersistentFlags().String("clientcas", client.DefaultMTLsOptions().ClientCAs, "clients certificates list. Aka certificate authority")
+	if err := viper.BindPFlag("default.port", cmd.PersistentFlags().Lookup("port")); err != nil {
+		return err
+	}
+	if err := viper.BindPFlag("default.address", cmd.PersistentFlags().Lookup("address")); err != nil {
+		return err
+	}
+	if err := viper.BindPFlag("default.mtls", cmd.PersistentFlags().Lookup("mtls")); err != nil {
+		return err
+	}
+	if err := viper.BindPFlag("default.servername", cmd.PersistentFlags().Lookup("servername")); err != nil {
+		return err
+	}
+	if err := viper.BindPFlag("default.certificate", cmd.PersistentFlags().Lookup("certificate")); err != nil {
+		return err
+	}
+	if err := viper.BindPFlag("default.pkey", cmd.PersistentFlags().Lookup("pkey")); err != nil {
+		return err
+	}
+	if err := viper.BindPFlag("default.clientcas", cmd.PersistentFlags().Lookup("clientcas")); err != nil {
+		return err
+	}
+	viper.SetDefault("default.port", gw.DefaultOptions().ImmudbPort)
+	viper.SetDefault("default.address", gw.DefaultOptions().ImmudbAddress)
+	viper.SetDefault("default.mtls", client.DefaultOptions().MTLs)
+	viper.SetDefault("default.servername", client.DefaultMTLsOptions().Servername)
+	viper.SetDefault("default.certificate", client.DefaultMTLsOptions().Certificate)
+	viper.SetDefault("default.pkey", client.DefaultMTLsOptions().Pkey)
+	viper.SetDefault("default.clientcas", client.DefaultMTLsOptions().ClientCAs)
+
+	return nil
 }
