+++ cmd/immuadmin/commands/user.go
@@ -30,13 +30,11 @@ func (cl *commandline) user(cmd *cobra.Command) {
 	ccmd := &cobra.Command{
 		Use:   "user username",
 		Short: "Perform various user operations: create, delete, change password",
-		Long: `Flags: --create, --delete, --password
-Examples:
-  - Create a new user 'username' (password will be generated by the server):
+		Example: `  Create a new user 'username' (password will be generated by the server):
     ./immuadmin user username --create
-  - Delete user 'username':
+  Delete user 'username':
     ./immuadmin user username --delete
-  - Change password for user 'username' (you will be prompted to confirm the old and the new passwords):
+  Change password for user 'username':
     ./immuadmin user username --password`,
 		Aliases:           []string{"u"},
 		PersistentPreRunE: cl.connect,
@@ -63,7 +61,8 @@ Examples:
 					c.QuitWithUserError(err)
 				}
 				fmt.Printf(
-					"User %s created with password %s\n",
+					"User created:\n---\nusername: %s\npassword: %s\n---\n"+
+						"IMPORTANT: This is the only time these credentials are shown, so make sure you remember them.\n",
 					string(response.User),
 					string(response.Password),
 				)
